test(sync-slim-agents): split the reconciliation cases into their own module
`test_sync_slim_agents.py` had reached 686 lines and carried a
`# taste-lint: ignore file-size` escape whose stated reason was a scoping
decision, not an engineering one. This removes the escape by moving the cases
that earned it, mirroring how `test_sync_slim_agents_contract.py` was split
from the same module.

`test_sync_slim_agents_transforms.py` takes what the reconciliation layer
decides and what `--check` reports: the `mcp__github__` strip, the
`mcp__serena__` passthrough, insert-only syncing, the two blocking shapes with
their file and line, the truncated excerpt, and the three-way drift counts.
The original module keeps the file layer: path safety, symlink refusal, atomic
write, mode preservation, exit codes, frontmatter preservation, encoding.

The loader and `tree` fixture are duplicated rather than shared. A pytest
fixture reaches another module only through a `conftest.py`, and this one is
too specific to belong to every test under `tests/build_scripts/`. The distinct
module name is the same isolation the contract module documents: loading here
must not displace a sibling's instance in `sys.modules`, or one module's
monkeypatched constants would leak into another's assertions.

61 tests pass across the three modules, unchanged in count and in what they
assert. Every file is now under the 500-line limit with no escape anywhere:
488 and 342 for the two build modules, 496, 213 and 267 for the three test
modules. `taste count ratchet: OK (count == baseline 575)`.

Refs #5282
